+++ ThirdPartyAdapters/bidmachine/bidmachine/src/test/kotlin/com/google/ads/mediation/bidmachine/BidMachineRewardedAdTest.kt
@@ -15,14 +15,17 @@
 package com.google.ads.mediation.bidmachine
 
 import android.content.Context
+import androidx.core.os.bundleOf
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.ads.mediation.adaptertestkit.AdErrorMatcher
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_BID_RESPONSE
+import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_PLACEMENT_ID
 import com.google.ads.mediation.adaptertestkit.createMediationRewardedAdConfiguration
 import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.ADAPTER_ERROR_DOMAIN
 import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.ERROR_CODE_AD_REQUEST_EXPIRED
 import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.ERROR_MSG_AD_REQUEST_EXPIRED
+import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.PLACEMENT_ID_KEY
 import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.SDK_ERROR_DOMAIN
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback
@@ -60,15 +63,37 @@ class BidMachineRewardedAdTest {
 
   @Before
   fun setUp() {
+    val serverParams = bundleOf(PLACEMENT_ID_KEY to TEST_PLACEMENT_ID)
     val adConfiguration =
-      createMediationRewardedAdConfiguration(context = context, bidResponse = TEST_BID_RESPONSE)
+      createMediationRewardedAdConfiguration(
+        context = context,
+        bidResponse = TEST_BID_RESPONSE,
+        serverParameters = serverParams,
+      )
     BidMachineRewardedAd.newInstance(adConfiguration, mockAdLoadCallback).onSuccess {
       bidMachineRewardedAd = it
     }
   }
 
   @Test
-  fun loadAd_invokesBidMachineRequest() {
+  fun loadWaterfallAd_invokesBidMachineRequest() {
+    val mockRewardedRequestBuilder =
+      mock<RewardedRequest.Builder> {
+        on { setPlacementId(eq(TEST_PLACEMENT_ID)) } doReturn it
+        on { setListener(any()) } doReturn it
+        on { build() } doReturn mockRewardedRequest
+      }
+    bidMachineRewardedAd.rewardedRequestBuilder = mockRewardedRequestBuilder
+
+    bidMachineRewardedAd.loadWaterfallAd(mockRewardedAd)
+
+    verify(mockRewardedRequestBuilder).setPlacementId(eq(TEST_PLACEMENT_ID))
+    verify(mockRewardedRequestBuilder).setListener(eq(bidMachineRewardedAd))
+    verify(mockRewardedRequest).request(eq(context))
+  }
+
+  @Test
+  fun loadRtbAd_invokesBidMachineRequest() {
     val mockRewardedRequestBuilder =
       mock<RewardedRequest.Builder> {
         on { setBidPayload(eq(TEST_BID_RESPONSE)) } doReturn it
@@ -77,7 +102,7 @@ class BidMachineRewardedAdTest {
       }
     bidMachineRewardedAd.rewardedRequestBuilder = mockRewardedRequestBuilder
 
-    bidMachineRewardedAd.loadAd(mockRewardedAd)
+    bidMachineRewardedAd.loadRtbAd(mockRewardedAd)
 
     verify(mockRewardedRequestBuilder).setBidPayload(eq(TEST_BID_RESPONSE))
     verify(mockRewardedRequestBuilder).setListener(eq(bidMachineRewardedAd))
@@ -86,7 +111,7 @@ class BidMachineRewardedAdTest {
 
   @Test
   fun showAd_invokesBidMachineShow() {
-    bidMachineRewardedAd.loadAd(mockRewardedAd)
+    bidMachineRewardedAd.loadRtbAd(mockRewardedAd)
 
     bidMachineRewardedAd.showAd(context)
 
@@ -95,7 +120,7 @@ class BidMachineRewardedAdTest {
 
   @Test
   fun onRequestSuccess_invokesBannerViewLoad() {
-    bidMachineRewardedAd.loadAd(mockRewardedAd)
+    bidMachineRewardedAd.loadRtbAd(mockRewardedAd)
 
     bidMachineRewardedAd.onRequestSuccess(mockRewardedRequest, mock())
 
@@ -108,7 +133,7 @@ class BidMachineRewardedAdTest {
     whenever(mockRewardedRequest.isExpired) doReturn true
     val expectedAdError =
       AdError(ERROR_CODE_AD_REQUEST_EXPIRED, ERROR_MSG_AD_REQUEST_EXPIRED, ADAPTER_ERROR_DOMAIN)
-    bidMachineRewardedAd.loadAd(mockRewardedAd)
+    bidMachineRewardedAd.loadRtbAd(mockRewardedAd)
 
     bidMachineRewardedAd.onRequestSuccess(mockRewardedRequest, mock())
 
@@ -150,7 +175,7 @@ class BidMachineRewardedAdTest {
   fun onAdLoadFailed_invokesOnFailure() {
     val bMError = BMError.AlreadyShown
     val expectedAdError = AdError(bMError.code, bMError.message, SDK_ERROR_DOMAIN)
-    bidMachineRewardedAd.loadAd(mockRewardedAd)
+    bidMachineRewardedAd.loadRtbAd(mockRewardedAd)
 
     bidMachineRewardedAd.onAdLoadFailed(mockRewardedAd, bMError)